+++ Model/TypeReflection/DbSelectColumnExtractor.php
@@ -7,7 +7,6 @@ use Magento\Framework\DB\Select;
 use function array_filter as filter;
 use function array_map as map;
 use function array_merge as merge;
-use function array_reduce as reduce;
 use function array_values as values;
 
 class DbSelectColumnExtractor
@@ -71,7 +70,10 @@ class DbSelectColumnExtractor
     {
         [$correlationName, $column, $alias] = $columnEntry;
         $tableName = $this->getRealTableNameIfAlias($select, $correlationName);
-        return $this->tableColumnExtractor->getColumnType($tableName, $column);
+
+        return is_string($column)
+            ? $this->tableColumnExtractor->getColumnType($tableName, $column)
+            : 'string'; // for Zend_Db_Expr instances default to string type
     }
 
     public function extractColumnValue(string $key, $object)